requirements.txt
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
DateTime==4.3
ecdsa==0.14.1
fastapi==0.65.1
greenlet==1.1.0
h11==0.12.0
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
python-jose==3.2.0
python-multipart==0.0.5
pytz==2021.1
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.15
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.13.4
zope.interface==5.4.0
